change some flags to subcommands
